User/robyn/branch (#8)
* Delete Linda written by Robyn.docx

* Linda (Business Owner) - Robyn


Persona
Linda – Business Owner
Linda is 35 years old and lives in a small town in MN. She and her husband just started their own brewery. While her husband oversees making the beer and finding the right ingredients, she oversees everything else. Linda is comfortable with schedules and puts everything in her calendar on her cell phone as things come up. When Linda is at the computer, she uses the website to see how much time she has between tasks. Although she is not computer savvy, she uses what is most convenient to her. Linda is in charge of creating community events to bring people together at the brewery and ensuring all the servers are scheduled as needed. She is also in charge of the local bands that are able to play at the brewery at least once a week. She wants everyone to be happy and to establish a good relation in the community.


User Stories

User Story 1 (Linda, business owner)
Linda has a hectic day and realizes since the grand opening several bands have reached out and had reviews scheduled on Monday. She doesn’t know how much time she has between the reviews to complete the regular daily requirements of the brewery. Although the brewery opens at 11 am all her band reviews are 11 am and after. She just made it to the brewery at 9 am and viewed her calendar. She views the countdown clock and sees she has exactly 1.50 hours until her first review. Without eating breakfast or starting her cleaning she jumps on the jobs. She gets most of the cleaning done and her next review is at 1:30 pm and she has another every 30 min thereafter. The first review lasted 45 min, and an employee was called in sick. Linda agreed to help where needed to keep the flow of business. She realizes with the countdown clock she now has 1.1 hours until the next review and it’s always busier after lunch. With the limited time she has she is able to call the bands and ask that they do the review in front of the guests. This works out perfectly, the reviews are given a 1 – 10 score and all the customers are happy to put their vote in. Linda was able to get through the reviews and made progress with the local bands.

User Story 2 (Linda, business owner)
Linda decided she would work from home today for the brewery and be on call when needed. She logs into the website and reviews her calendar. She sees there is no countdown, but the business is in the summer parade in just 3 days! She forgot to put a countdown on the event. Now that she has added the countdown, she can see that she has 2 days and 12 hours left. She scrambles to the store to get things as needed and rushes her online delivery orders. The countdown helps, and when she doesn’t add it to her calendar events, she forgets about them.